Update deploy script to make tags work
Since I have a `tags` folder in my global .gitignore file the `tags`
folder in the `public` folder of the website was not being committed to
the repository, thus the tags were not working on the live version.

This commit updates the deploy script to forcefully (-f) add all the
files inside the public folder, which will add all nested folders inside
`public`, making the tags system work again.