Make links visible in black and white
This commit updates the `a` styles so that if someone is looking at a
blogpost in black and white on their phone they can still discern where
are the links in the middle of the text.

It also includes a change in the config.toml file so as to change the
website's URL to use https.